# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vaishnavi Residency Hotel Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: getReservations
      tags: [reservations]
      summary: List all reservations
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      summary: Create a reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      summary: Get a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    patch:
      operationId: updateReservation
      tags: [reservations]
      summary: Update a reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Reservation updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"

  # PAYMENTS
...
  /reports/revenue:
    get:
      operationId: getRevenueReport
      tags: [reports]
      summary: Get revenue report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Revenue report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueReport"
  /reports/occupancy:
    get:
      operationId: getOccupancyReport
      tags: [reports]
      summary: Get occupancy report
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [today, week, month, year]
      responses:
        "200":
          description: Occupancy report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OccupancyReport"
...
components:
  schemas:
...
    Reservation:
      type: object
      required: [id, reservationNumber, guestName, roomId, checkInDate, checkOutDate, status]
      properties:
        id:
          type: integer
        reservationNumber:
          type: string
        guestName:
          type: string
        mobile:
          type: string
        roomId:
          type: integer
        roomNumber:
          type: ["string", "null"]
        checkInDate:
          type: string
        checkOutDate:
          type: string
        advance:
          type: number
        status:
          type: string
          enum: [reserved, confirmed, cancelled]
        remarks:
          type: ["string", "null"]
        createdAt:
          type: string
    ReservationInput:
      type: object
      required: [guestName, mobile, roomId, checkInDate, checkOutDate]
      properties:
        guestName:
          type: string
        mobile:
          type: string
        roomId:
          type: integer
        checkInDate:
          type: string
        checkOutDate:
          type: string
        advance:
          type: number
        remarks:
          type: string
    ReservationUpdate:
      type: object
      properties:
        guestName:
          type: string
        mobile:
          type: string
        roomId:
          type: integer
        checkInDate:
          type: string
        checkOutDate:
          type: string
        advance:
          type: number
        status:
          type: string
          enum: [reserved, confirmed, cancelled]
        remarks:
          type: string
...
    RevenueReport:
      type: object
      required: [period, totalRevenue, totalGuests, data]
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        totalGuests:
          type: integer
        pendingBalance:
          type: number
        data:
          type: array
          items:
            $ref: "#/components/schemas/RevenueDataPoint"
    RevenueDataPoint:
      type: object
      required: [label, revenue, guests]
      properties:
        label:
          type: string
        revenue:
          type: number
        guests:
          type: integer
    OccupancyReport:
      type: object
      required: [period, occupancyRate, totalRooms, occupiedRooms, data]
      properties:
        period:
          type: string
        occupancyRate:
          type: number
        totalRooms:
          type: integer
        occupiedRooms:
          type: integer
        topRooms:
          type: array
          items:
            $ref: "#/components/schemas/RoomOccupancyStat"
        data:
          type: array
          items:
            $ref: "#/components/schemas/OccupancyDataPoint"
    RoomOccupancyStat:
      type: object
      required: [roomNumber, bookings]
      properties:
        roomNumber:
          type: string
        bookings:
          type: integer
    OccupancyDataPoint:
      type: object
      required: [label, occupancyRate]
      properties:
        label:
          type: string
        occupancyRate:
          type: number
...